requirements.txt
Django>=2.0,<3.0
psycopg2>=2.7,<3.0
djangorestframework
django-filter
requests
django-cors-headers
django-environ
django-extensions
django-debug-toolbar
Werkzeug
ipython